+++ changelog.md
@@ -5,6 +5,7 @@
 * Recall idol 
   * Now has a decent texture (a golden bed basically)
   * Now has "animation" and particles
+* The enchanting table ui has been enlarged a little bit
 * Animals and sea creatures are now reduced in Autumn
 * Bone meal now has 40% and 80% chance to fail to grow plants in autumn and winter respectively
 * Added the `iguanatweaks:doSeasonGrassGrowDeath` gamerule to prevent grass from growing and dying based off seasons